AC2.0.24b
git-svn-id: https://arducopter.googlecode.com/svn/trunk@2567 f9c3cf11-9bcb-44bc-f272-b75c42450872

+++ ArduCopterMega/Mavlink_Common.h
@@ -221,7 +221,7 @@ void mavlink_send_message(mavlink_channel_t chan, uint8_t id, uint32_t param, ui
 					(float)airspeed / 100.0,
 					(float)g_gps->ground_speed / 100.0,
 					(dcm.yaw_sensor / 100) % 360,
-					(int)g.rc_3.servo_out,
+					g.rc_3.servo_out,
 					current_loc.alt / 100.0,
 					climb_rate);
 			break;
